backend/requirements.txt
debugpy==1.8.7
bcrypt==4.3.0
flask==3.0.0
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
python-json-logger==2.0.7
psycopg2-binary==2.9.9
pydantic==2.5.0
email-validator==2.1.0
pytest==7.4.3
pytest-flask==1.3.0
